requirements.txt
streamlit
wordcloud
PyPDF2
python-docx
pillow
pandas
